+++ CHANGELOG.md
@@ -1,8 +1,7 @@
 ## 0.6.0 (unreleased)
 
 - Added support for generated columns
-- Added support for extended statistics
-- Added compression to `prep`
+- Added compression and extended statistics to `prep`
 - Dropped support for Ruby < 2.7
 
 ## 0.5.0 (2023-01-29)